pyproject.toml
[project]
name = "career-sync-ai"
version = "0.1.0"
description = "AI-powered resume tailoring with LangGraph and Mistral 7B"
requires-python = ">=3.12"
authors = [
    {name = "QubridAI", email = "dev@qubrid.com"}
]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "streamlit>=1.30.0",
    "langgraph>=0.0.40",
    "langchain-core>=0.1.0",
    "openai>=1.10.0",
    "python-dotenv>=1.0.0",
    "pypdf2>=3.0.0",
    "reportlab>=4.0.0",
    "markdown>=3.5.0",
    "pydantic>=2.5.0",
    "typing-extensions>=4.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "backend", "frontend"]

[dependency-groups]
dev = []
